requirements.txt
psycopg2==2.9.6
PyYAML==6.0
watchdog==3.0.0
